+++ PolygonChecker/main.h
@@ -2,4 +2,4 @@
 
 void printWelcome(); 
 int printShapeMenu();
-int* getTriangleSides(int* triangleSides);
+double* getTriangleSides(double* triangleSides);
